something crashing i must build that crash reporter

+++ FortGen/Engine/Source/Runtime/Core/Public/Containers/PropertyMarcos.h
@@ -6,7 +6,7 @@
     void Set##Name(Type Value) { *reinterpret_cast<Type*>(reinterpret_cast<uintptr_t>(this) + Offset) = Value; }
 
 #define DEFINE_STRUCTSIZE(Size) \
-    static int32_t GetStructSize() { return Size; }
+    static size_t GetStructSize() { return Size; }
 
 #define DEFINE_DATAINDEX(Name, Offset, Type, Structsize) \
-    Type Get##Name(int32_t Index) const { return *reinterpret_cast<Type*>(*reinterpret_cast<uintptr_t*>(reinterpret_cast<uintptr_t>(this) + Offset) + Structsize * Index); }
+    Type Get##Name(size_t Index) const { return *reinterpret_cast<Type*>(*reinterpret_cast<uintptr_t*>(reinterpret_cast<uintptr_t>(this) + Offset) + Structsize * Index); }
